Backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
python-multipart>=0.0.6
pillow>=10.2.0
numpy>=1.26.0
pydantic>=2.5.0
